Django_API/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
audioread==3.0.1
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.1
cryptography==44.0.3
cycler==0.12.1
dash==3.0.4
debugpy==1.8.14
decorator==5.2.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
flasgger==0.9.7.1
Flask==3.1.1
Flask-Cors==4.0.0
fonttools==4.57.0
fsspec==2025.5.1
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.6
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.14
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mistune==3.1.3
mpmath==1.3.0
msgpack==1.1.0
narwhals==1.37.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
numba==0.61.2
numpy==2.2.4
open3d==0.19.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.7
plotly==6.0.1
pluggy==1.5.0
pooch==1.8.2
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyassimp==5.2.5
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rpds-py==0.25.1
scikit-learn==1.7.0
scipy==1.15.2
seaborn==0.13.2
setuptools==80.3.0
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.1
torchvision==0.22.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trimesh==4.6.8
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.158
ultralytics-thop==2.0.14
urllib3==2.4.0
vedo==2025.5.3
vtk==9.4.2
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.14
zipp==3.21.0
